Separate the MySql tests from the backend independent ones
